+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
   - Fixes entrypoint assertion error ([dfm](https://github.com/dfm)), refs #5
+  - Fixes parallel mode ([johanneswilm](https://github.com/johanneswilm)), refs #8
+  - Improves exception logging
 
 ## [v20200413]
   - Made `github-token` parameter optional